move tests into tests directory
